this branch 
```bash
Gnuplot not found, using plotters backend
recursive_parser/small(7K)                                                                            
                        time:   [42.298 us 42.547 us 42.754 us]
                        change: [-20.656% -19.481% -18.441%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  11 (11.00%) low mild

Benchmarking recursive_parser/large(201K): Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 7.2s, enable flat sampling, or reduce sample count to 50.
recursive_parser/large(201K)                                                                             
                        time:   [1.0933 ms 1.1002 ms 1.1058 ms]
                        change: [-19.383% -17.867% -16.332%] (p = 0.00 < 0.05)
                        Performance has improved.

```

main branch 
```bash
recursive_parser/small(7K)                                                                            
                        time:   [51.944 us 52.056 us 52.149 us]
                        change: [+21.004% +22.630% +24.207%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 11 outliers among 100 measurements (11.00%)
  10 (10.00%) low severe
  1 (1.00%) low mild

Benchmarking recursive_parser/large(201K): Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 8.3s, enable flat sampling, or reduce sample count to 50.
recursive_parser/large(201K)                                                                             
                        time:   [1.3027 ms 1.3081 ms 1.3125 ms]
                        change: [+18.758% +20.880% +23.165%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 9 outliers among 100 measurements (9.00%)
```