@IWANABETHATGUY 2021-12-02
this would also helpful when we try to implement a LL(2) parser (related to #104 ) without introduce too much overhead.