@justjavac 2021-12-03
use

```bash
cargo build --release --locked --bin "benchmark*"
cargo run --release --bin benchmark
```

it is slower

![image](https://user-images.githubusercontent.com/359395/144525970-278a7021-0777-4c8d-b540-1f0db8bde3f4.png)